pyproject.toml
[project]
name = "mcp-simple-slackbot"
version = "0.1.0"
description = "A simple Slack bot using the Model Context Protocol (MCP)"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "sooperset", email = "soomiles.dev@gmail.com" }]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "slack_bolt>=1.18.0",
    "slack_sdk>=3.21.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
    "httpx>=0.24.1",
    "aiohttp>=3.11.13",
    "uvicorn>=0.23.2",
    "requests>=2.32.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
include = ["mcp_simple_slackbot"]
venvPath = "."
venv = "venv"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
